Break out XivChatType Extensions and InfoAttribute
